Use wide_sampling in fx2lafw only if necessary.
This changes the fx2lsfw code to only enable the wide_sampling (16Bit)
method, if at least one of the higher eight channels is enabled. This
has the benefit, that we can use the doubled samplerate on 16Bit LA
devices if we use only the first eight channels.

The wide sampling is also enabled if we have one or more analog
channels.

Signed-off-by: Jan Losinski <losinski@wh2.tu-dresden.de>